summary.py
#!/usr/bin/env python3
"""
Project Summary and Test Suite
Shows all bot capabilities and runs basic tests
"""

import sys
import os

def print_header(text):
    print("\n" + "="*70)
    print(f"  {text}")
    print("="*70 + "\n")

def print_section(emoji, title, items):
    print(f"\n{emoji} {title}")
    print("-" * 70)
    for item in items:
        print(f"  ✓ {item}")

def main():
    print_header("🤖 CRYPTO TRADING BOT - PROJECT SUMMARY")
    
    print("""
    ╔══════════════════════════════════════════════════════════════╗
    ║                                                              ║
    ║         PROFESSIONAL CRYPTOCURRENCY TRADING BOT              ║
    ║                                                              ║
    ║  Built with: Python, CCXT, Technical Analysis & ML           ║
    ║  Status: Ready for Testing & Deployment                      ║
    ║                                                              ║
    ╚══════════════════════════════════════════════════════════════╝
    """)
    
    print_section("✨", "FEATURES", [
        "Multi-exchange support (Binance, Coinbase, Kraken, etc.)",
        "Advanced technical analysis (RSI, MACD, EMA, Bollinger Bands)",
        "Comprehensive backtesting engine with visualization",
        "Real-time market monitoring and signal generation",
        "Paper trading mode for risk-free practice",
        "Automated stop-loss and take-profit management",
        "Strategy optimization and parameter tuning",
        "Trade history tracking and performance analytics"
    ])
    
    print_section("📊", "TECHNICAL INDICATORS", [
        "RSI (Relative Strength Index) - Momentum indicator",
        "MACD (Moving Average Convergence Divergence) - Trend following",
        "EMA (Exponential Moving Average) - Price smoothing",
        "Bollinger Bands - Volatility indicator",
        "ATR (Average True Range) - Volatility measurement",
        "Multi-timeframe analysis support"
    ])
    
    print_section("🛡️", "RISK MANAGEMENT", [
        "Configurable stop-loss percentage (default: 2%)",
        "Configurable take-profit targets (default: 5%)",
        "Position sizing controls (default: 95% of capital)",
        "Paper trading mode for safe testing",
        "API rate limiting to prevent overuse",
        "Error handling and graceful recovery"
    ])
    
    print_section("🚀", "AVAILABLE COMMANDS", [
        "python demo.py          - Run simulation with fake data (SAFE)",
        "python backtest.py      - Test on historical data",
        "python monitor.py       - Real-time market monitoring",
        "python bot.py           - Live/Paper trading bot",
        "python optimize.py      - Find optimal parameters",
        "python config.py        - Interactive configuration",
        "./start.sh              - Quick start menu"
    ])
    
    print_section("📁", "PROJECT FILES", [
        "bot.py          - Main trading bot (400+ lines)",
        "backtest.py     - Backtesting module (300+ lines)",
        "monitor.py      - Market monitor (100+ lines)",
        "demo.py         - Demo simulator (150+ lines)",
        "optimize.py     - Strategy optimizer (150+ lines)",
        "config.py       - Configuration wizard (150+ lines)",
        "requirements.txt - Python dependencies",
        ".env            - Configuration file"
    ])
    
    print_header("🧪 RUNNING BASIC TESTS")
    
    # Test 1: Import check
    print("Test 1: Checking dependencies...")
    try:
        import ccxt
        import pandas
        import numpy
        import ta
        from dotenv import load_dotenv
        print("  ✅ All dependencies installed correctly")
    except ImportError as e:
        print(f"  ❌ Missing dependency: {e}")
        return
    
    # Test 2: Configuration check
    print("\nTest 2: Checking configuration...")
    if os.path.exists('.env'):
        print("  ✅ .env configuration file exists")
    else:
        print("  ⚠️  .env not found (will use defaults)")
    
    # Test 3: Module imports
    print("\nTest 3: Testing bot modules...")
    try:
        from bot import CryptoTradingBot
        from backtest import Backtester
        print("  ✅ Bot modules load successfully")
    except Exception as e:
        print(f"  ❌ Module error: {e}")
        return
    
    # Test 4: Demo run
    print("\nTest 4: Running quick demo...")
    try:
        from demo import DemoBot
        demo = DemoBot()
        print("  ✅ Demo bot initialized successfully")
        print("  💡 Run 'python demo.py' for full simulation")
    except Exception as e:
        print(f"  ❌ Demo error: {e}")
    
    print_header("📋 NEXT STEPS")
    
    print("""
    🎯 RECOMMENDED WORKFLOW:
    
    1. 📖 READ:      Check README.md and QUICKSTART.md
    2. 🎮 DEMO:      Run 'python demo.py' to see it work
    3. ⚙️  CONFIG:    Run 'python config.py' to set up
    4. 🔑 API KEYS:  Get keys from your exchange
    5. 🔄 BACKTEST:  Run 'python backtest.py' to test
    6. 📊 MONITOR:   Run 'python monitor.py' to watch
    7. 📝 PAPER:     Run bot with PAPER_TRADING=true
    8. ⚠️  LIVE:      Only after extensive testing!
    
    ⚠️  IMPORTANT WARNINGS:
    
    • Cryptocurrency trading is HIGHLY RISKY
    • Never invest more than you can afford to lose
    • Past performance does NOT guarantee future results
    • Start with PAPER TRADING (simulated trades)
    • Test extensively before using real money
    • Markets are extremely volatile - losses can be severe
    • This bot is for EDUCATIONAL purposes
    • The developers are NOT responsible for any losses
    
    📚 RESOURCES:
    
    • Full documentation:  README.md
    • Quick start guide:   QUICKSTART.md
    • Configuration help:  python config.py
    • Demo simulation:     python demo.py
    
    🤝 SUPPORT:
    
    • Check documentation for troubleshooting
    • Review .env.example for configuration options
    • Test with demo mode first (safest option)
    • Start with small amounts when live trading
    
    """)
    
    print_header("✅ SETUP COMPLETE - READY TO USE")
    
    print("\nQuick Start:")
    print("  $ python demo.py          # Safe demo with fake data")
    print("  $ python config.py        # Configure your settings")
    print("  $ python backtest.py      # Test on real historical data")
    print("\n")


if __name__ == "__main__":
    main()
